### 2019-11-28 #1919 @Kevin-Mattheus-Moerman
@openjournals/dev can you see what is wrong here? The affiliation index does use quotes.
