### 2019-11-30 #1919 @xuanxu
@Kevin-Mattheus-Moerman, @raullaasner: There was an error with the indexing of the affiliations, [this PR](https://github.com/HybriD3-database/MatD3/pull/2) should fix the compiling error.
