### 2019-12-01 #1919 @Kevin-Mattheus-Moerman
@raullaasner can you also work on those missing DOI's? You can run `@whedon check references` and `@whedon generate pdf` here too to check things.
