### 2019-12-01 bd01cfa
Include DOIs in paper references
