+++ paper/refs.bib
@@ -8,6 +8,7 @@
   journal = 	 {Scientific Data},
   year = 	 2019,
   volume = 	 6,
+  doi =          "10.1038/sdata.2019.2",
   pages = 	 190002}
 
 @Article{jain13,
@@ -22,6 +23,7 @@
   journal = 	 {APL Materials},
   year = 	 2013,
   volume = 	 1,
+  doi =          "10.1063/1.4812323",
   pages = 	 011002}
 
 @misc{nomad,
@@ -38,6 +40,7 @@
   journal = 	 {JOM},
   year = 	 2016,
   volume = 	 68,
+  doi =          "10.1007/s11837-016-2001-3",
   pages = 	 {2045--2052}}
 
 @misc{springermaterials,
@@ -53,6 +56,7 @@
   journal = 	 {Comput. Mater. Sci.},
   year = 	 2012,
   volume = 	 58,
+  doi =          "10.1016/j.commatsci.2012.02.005",
   pages = 	 {218--226}}
 
 @Article{saal13,
@@ -64,4 +68,5 @@
   journal = 	 {JOM},
   year = 	 2013,
   volume = 	 65,
+  doi =          "10.1007/s11837-013-0755-4",
   pages = 	 {1501}}
