### 2019-12-01 #1919 @raullaasner
@Kevin-Mattheus-Moerman I added DOIs to the paper references except for the SpringerMaterials citation, which is only a web link. It seems there is no better way to cite it at this point.
