### 2019-12-01 #1919 @raullaasner
Regarding the proof, would it possible have "3" as a superscript in MatD<sup>3</sup>? In the paper, I had written `MatD<sup>3</sup>` everywhere but this formatting seems not to be present in the proof.
