### 2019-12-01 #1919 @Kevin-Mattheus-Moerman
@openjournals/dev do you know how to use superscript in the text as the author requests?

> Regarding the proof, would it possible have "3" as a superscript in MatD3? In the paper, I had written `MatD<sup>3</sup>` everywhere but this formatting seems not to be present in the proof.
