### 2019-12-05 52bed50
[security fix] Bump Django to 2.2.8
