### 2019-12-05 0bbf9b3
Bump django from 2.2.8 to 3.0
Bumps [django](https://github.com/django/django) from 2.2.8 to 3.0.
- [Release notes](https://github.com/django/django/releases)
- [Commits](https://github.com/django/django/compare/2.2.8...3.0)

Signed-off-by: dependabot[bot] <support@github.com>
