+++ requirements.txt
@@ -1,4 +1,4 @@
-Django==2.2.8
+Django==3.0
 matplotlib==3.0.2
 mysqlclient==1.4.2
 numpy==1.16.1
