### 2019-12-06 #1919 @majensen
Hey @pythonpanda2, would you be willing to review this submission for [JOSS](https://joss.theoj.org)? Please let us know in this thread. Thanks for the consideration!
