### 2019-12-09 #1919 @arfon
> Regarding the proof, would it possible have "3" as a superscript in MatD3? In the paper, I had written MatD<sup>3</sup> everywhere but this formatting seems not to be present in the proof.

@raullaasner you can do `MatD^3^` which should work.
