### 2019-12-09 d9d7c83
Replace <sup>3</sup> with ^3^ in paper
