+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'MatD<sup>3</sup>: A Database and Online Presentation Package for Research Data Supporting Materials Discovery, Design, and Dissemination'
+title: 'MatD^3^: A Database and Online Presentation Package for Research Data Supporting Materials Discovery, Design, and Dissemination'
 tags:
   - Python
   - Django
@@ -48,17 +48,17 @@ bibliography: refs.bib
 
 The discovery of new materials as well as the determination of a vast set of materials properties for science and technology is a fast growing field of research, with contributions from many groups worldwide. Materials data from individual research groups is traditionally disseminated by means of loosely interconnected, peer-reviewed publications. Several data-centric efforts such as the Materials Project [@jain13], Nomad [@nomad], Aflow [@curtarolo12], The Open Quantum Materials Database [@saal13], or the Materials Data Facility [@blaiszik16] are making large segments of materials data publicly available, using a central repository. Because of the complexity and heterogeneity of materials science data, and the difficulty, in some cases, to assign given data to specific databases, researchers may find preferable to make data available in a distributed manner, on an individual basis, using locally hosted and controlled solutions. Of course distributed solutions do not exclude that data made available individually may also be contributed to large, centralized databases.
 
-The package presented here (MatD<sup>3</sup>) is intended to be a simple solution to make diverse datasets available individually and rapidly for reproducibility or other purposes. Furthermore, traditional data curation efforts such as SpringerMaterials [@springermaterials] or topical reviews have long supported the task of summarizing diverse sets of materials data from the community, but data in curated, reusable collections are not yet available for many classes of materials and material properties. Small and large research groups are struggling to find and access scientific data presented in scientific papers due to the lack of established queriable web interfaces to heterogenous sets of data.
+The package presented here (MatD^3^) is intended to be a simple solution to make diverse datasets available individually and rapidly for reproducibility or other purposes. Furthermore, traditional data curation efforts such as SpringerMaterials [@springermaterials] or topical reviews have long supported the task of summarizing diverse sets of materials data from the community, but data in curated, reusable collections are not yet available for many classes of materials and material properties. Small and large research groups are struggling to find and access scientific data presented in scientific papers due to the lack of established queriable web interfaces to heterogenous sets of data.
 
-MatD<sup>3</sup> is an open-source, dedicated database and web application framework designed to store, curate and disseminate experimental and theoretical materials data generated by individual research groups or research consortia. A research group can set up its own instance of MatD<sup>3</sup> and publish scientific results or simply use an existing online MatD<sup>3</sup> instance. Disseminating research data in this form enables broader access, reproducibility, and repurposing of scientific products. MatD<sup>3</sup> is a general purpose database that does not focus on any specific level of theory or experimental method. Instead, the focus is on storing and making accessible the data and making it straightforward to curate them.
+MatD^3^ is an open-source, dedicated database and web application framework designed to store, curate and disseminate experimental and theoretical materials data generated by individual research groups or research consortia. A research group can set up its own instance of MatD^3^ and publish scientific results or simply use an existing online MatD^3^ instance. Disseminating research data in this form enables broader access, reproducibility, and repurposing of scientific products. MatD^3^ is a general purpose database that does not focus on any specific level of theory or experimental method. Instead, the focus is on storing and making accessible the data and making it straightforward to curate them.
 
[…]
