### 2019-12-09 d2398e6
Following 0bbf9b35, also bump some other packages
Django 3.0 removes django.utils.six, which is a requirement for
nested-admin and the REST framework. Bump versions of those packages
to fix the dependency issue.
