+++ requirements.txt
@@ -4,12 +4,11 @@ mysqlclient==1.4.2
 numpy==1.16.1
 Pillow==6.2.0
 python-dateutil==2.8.0
-six==1.11.0
 sentry-sdk==0.7.3
 raven==6.10.0
 python-decouple==3.1
-django-nested-admin==3.2.3
-djangorestframework==3.9.2
+django-nested-admin==3.2.4
+djangorestframework==3.10.3
 coverage==4.5.3
 selenium==3.141.0
 requests==2.22.0
