### 2019-12-09 #1919 @raullaasner
@arfon Thanks! I made the change and can confirm that it produces the desired output.
