### 2019-12-10 #1919 @majensen
Dear @computron - would you be willing to review the [paper](https://github.com/openjournals/joss-papers/blob/joss.01919/joss.01919/10.21105.joss.01919.pdf) in this issue for [Journal of Open Source Software](https://theoj.joss.org)? If you have any questions, please feel free to [email me](mailto:maj.fortinbras@gmail.com) . Thanks for your consideration! --@majensen
