### 2019-12-10 #1919 @pythonpanda2
@majensen

I was reviewing the  JOSS Conflict of Interest (COI) Policy.  I regret to inform you that my current employment at ANL and an ongoing collaboration with submitter research group might lead to COI.

I would like to nominate @aykol for review.
