### 2019-12-10 #1919 @majensen
Thanks for your diligence on that issue @pythonpanda2 -- @aykol are you willing to review this submission?
