### 2019-12-10 #1919 @majensen
Hi @dgasmith - would you be willing to review this submission to [JOSS](https://joss.theoj.org)?
