### 2019-12-10 #1919 @mkhorton
I'd be happy to take a look at this, I can't make a strong commitment on time however.
