### 2019-12-11 #1919 @majensen
Thanks for your response @computron -
@mkhorton - thx - what would be a workable timeline for you?
