### 2019-12-12 #1945 @whedon
[REVIEW]: MatD3: A Database and Online Presentation Package for Research Data Supporting Materials Discovery, Design, and Dissemination
**Submitting author:** @raullaasner (<a href="http://orcid.org/0000-0002-9714-1189">Raul Laasner</a>)
**Repository:** <a href="https://github.com/HybriD3-database/MatD3" target ="_blank">https://github.com/HybriD3-database/MatD3</a>
**Version:** v1
