### 2019-12-14 #1945 @mkhorton
@raullaasner what's the easiest way to import some example data for testing?
