### 2019-12-16 8cf530e
Fix reference and system forms on re-submit
When the "Add Data" form is reloaded because of errors, the reference
and system forms were previously not correctly displayed.
