### 2019-12-16 #1945 @raullaasner
@mkhorton The easiest way is to login on the website as the superuser (or create a regular user and use that) and click on the "Add Data" button. This presents you with a lot fields describing the data set, but only a few are mandatory. Click on "Submit" at the bottom to see which ones are mandatory and fill those in with some made up data.

Other than that, there is no import feature, i.e., the data needs to be manually inserted. If such a feature is necessary, I could implemented it in a day or two.
