### 2019-12-16 #1945 @mkhorton
Ok, that should be sufficient, thank you. I just need it to test functionality :-) Reading the docs made it sound like it needed to be connected to a Qresp backend.
