### 2020-01-03 230234c
Create pythonpackage.yml
Exclude Selenium test cases from the CI.
