### 2020-01-06 f67f40a
Add CI badge
Fixes #6
