### 2020-01-06 #1945 @dgasmith
My requested changes have been integrated! I believe I have completed the review process.
