### 2020-01-10 #1945 @mkhorton
I've also completed my review on this code and it's ready to go. Overall, this is a nice example of using the Django framework to create an interactive web app for exploring a materials database.
