### 2020-01-12 #1945 @majensen
@raullaasner the proof looks good. Can you please create an archive of the software repository using [Zenodo](https://zenodo.org) or similar? [This comment](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729) gives a quick overview for creating one on Zenodo.

When you have done that, please report the archive DOI back to us here -- thanks!
