### 2020-01-15 #1945 @majensen
👋 @raullaasner  - this paper is nearing acceptance; can you please review the [previous comment](https://github.com/openjournals/joss-reviews/issues/1945#issuecomment-573376626) and provide a software archive? thanks
