### 2020-01-16 #1945 @majensen
Hello @openjournals/joss-eics - the reviewers have recommended acceptance of #1945. @raullaasner has provided the archive, which checks out for title and authors, and whedon has done his duty. Thanks!
