### 2020-01-16 #1945 @kyleniemeyer
Hey @majensen, just FYI I was leaving the issue open until the PDF builds/resolves on the article page—you can generally leave that to the handling AEiC
